git-svn-id: https://opengl-scut.googlecode.com/svn/trunk@733 81c3a017-244d-672b-1dc6-c8d6962b81c2

+++ Code/Scar/Scar/RobotShip.h
@@ -29,7 +29,7 @@ public:
 private:
 	RobotState		State;
 	bool			fireOnce;
-	f32				TimePoint;
+	u32				TimePoint;
 
 };
 #endif // RobotShip_h__
